fix test

+++ syftbox/client/client2.py
@@ -261,7 +261,7 @@ def prompt_delete_old_data_dir(data_dir: Path) -> bool:
     return Confirm.ask(msg)
 
 
-def run_migration(config: SyftClientConfig):
+def run_migration(config: SyftClientConfig, prompt_remove=True):
     # first run config migration
     config.migrate()
 
@@ -275,7 +275,7 @@ def run_migration(config: SyftClientConfig):
 
     # Option 1: if outdated, completely remove the existing syftbox folder and start fresh
     if new_ws.data_dir.exists():
-        if has_old_syftbox_version(new_ws.data_dir) and prompt_delete_old_data_dir(new_ws.data_dir):
+        if has_old_syftbox_version(new_ws.data_dir) and prompt_delete_old_data_dir(new_ws.data_dir) and prompt_remove:
             logger.info("Removing old syftbox folder")
             shutil.rmtree(str(old_sync_folder))
             return
